修正页面样式，作者头像可选，每篇文章浏览量显示

+++ Test.DAL/Entitys/VisitAmount.cs
@@ -14,7 +14,7 @@ namespace DAL.Entitys
         /// 主键
         /// </summary>
         [Key]
-        public int id { set; get; }
+        public long id { set; get; }
 
         /// <summary>
         /// IP地址
